requirements.txt
fastapi
streamlit
scikit-learn
pandas
matplotlib
pillow